consolidate icon styles, share group
