+++ src/components/styles/Quote.js
@@ -1,7 +1,6 @@
 import styled from 'styled-components'
+
 const Quote = styled.div`
-  /* margin-top: -1.2rem; */
-  /* margin-bottom: 1.5rem; */
   margin: 3rem auto;
   text-align: center;
   font-size: 0.9rem;
