+++ src/pages/blog.js
@@ -1,11 +1,9 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import { graphql as gql, Link } from 'gatsby'
-import styled from 'styled-components'
 import chunk from 'lodash/chunk'
 import SEO from 'components/seo'
 import Layout from 'components/Skeleton'
-import theme from 'components/styles/theme'
 import Quote from 'components/styles/Quote'
 import { StyledPostLink, StyledTagList, StyledTag } from 'components/styles/Tags.css'
 import { StyledLoadButton } from 'components/styles/buttons'
